+++ src/main/java/conspire/relics/Flyswatter.java
@@ -2,42 +2,71 @@ package conspire.relics;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.HashSet;
 
 import com.megacrit.cardcrawl.cards.AbstractCard;
 import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.helpers.CardLibrary;
 import com.megacrit.cardcrawl.relics.AbstractRelic;
 
 public class Flyswatter extends AbstractConspireRelic {
     public static final String ID = "conspire:Flyswatter";
-    boolean used = false;
+    private boolean used = false;
+    private HashSet<String> skippedCards = new HashSet<>();
+    private int numCardChoicesProcessed = 0;
 
     public Flyswatter() {
         super(ID, AbstractRelic.RelicTier.UNCOMMON, AbstractRelic.LandingSound.FLAT);
     }
 
-    public static boolean checkSkipped(AbstractCard card) {
-        if (!AbstractDungeon.player.hasRelic(ID)) return false;
+    // get a card that was not previously skipped (if you have Flyswatter)
+    public static AbstractCard getCard(AbstractCard.CardRarity rarity, boolean anyclass) {
+        AbstractCard card = anyclass ? CardLibrary.getAnyColorCard(rarity) : AbstractDungeon.getCard(rarity);
+        // Check if card was skipped previously. Do this a limited number of times
+        for (int i = 0 ; i < 4 && checkSkippedStatic(card) ; ++i) {
+            card = anyclass ? CardLibrary.getAnyColorCard(rarity) : AbstractDungeon.getCard(rarity);
+        }
+        return card;
+    }
+
+    public static boolean checkSkippedStatic(AbstractCard card) {
+        AbstractRelic relic = AbstractDungeon.player.getRelic(ID);
+        if (relic != null) {
+            return ((Flyswatter)relic).checkSkipped(card);
+        } else {
+            return false;
+        }
+    }
+
+    private boolean checkSkipped(AbstractCard card) {
         // Has this card previously been skipped?
-        // This is a bit hacky, but we can use the metric data
-        // see CardRewardScreen.recordMetrics
+        updateSkippedCards();
         String metricID = card.getMetricID();
-        for (@SuppressWarnings("rawtypes") HashMap card_choice_ : CardCrawlGame.metricData.card_choices) {
+        if (skippedCards.contains(metricID)) {
+            this.used = true;
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    private void updateSkippedCards() {
+        // This is a bit hacky, but we can use the metric data
+        //  see CardRewardScreen.recordMetrics
+        // Note 2: We use a HashSet to cache the set of skipped cards, this way we don't loop over the metric data each time
+        for (int i = numCardChoicesProcessed ; i < CardCrawlGame.metricData.card_choices.size() ; ++i) {
             @SuppressWarnings("unchecked")
-            HashMap<String,Object> choice = (HashMap<String,Object>)card_choice_;
+            HashMap<String,Object> choice = (HashMap<String,Object>)CardCrawlGame.metricData.card_choices.get(i);
             Object picked = choice.get("picked");
             if (picked != null && (picked.equals("SKIP") || picked.equals("Singing Bowl"))) {
                 // player picked skip or singing bowl
                 @SuppressWarnings("unchecked")
                 ArrayList<String> notpicked = (ArrayList<String>)choice.get("not_picked");
-                if (notpicked.contains(metricID)) {
-                    Flyswatter relic = (Flyswatter)AbstractDungeon.player.getRelic(Flyswatter.ID);
-                    relic.used = true;
-                    return true;
-                }
+                skippedCards.addAll(notpicked);
             }
         }
-        return false;
+        numCardChoicesProcessed = CardCrawlGame.metricData.card_choices.size();
     }
 
     public void flashIfUsed() {