+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### HEAD
+* Added relics: Gift Box
+
 ### v1.2.2
 * Update for patch 46
 * Fix Infinite Journal: copied cards didn't inherit magicNumber of multiply upgraded cards